REFACTOR: Relocate HttpRequestUtils from http to web package
