+++ src/main/java/me/tiary/common/filter/AccessLoggingFilter.java
@@ -5,7 +5,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import me.tiary.common.util.http.HttpRequestUtils;
+import me.tiary.common.util.web.HttpRequestUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
