+++ src/main/java/me/tiary/common/util/web/HttpRequestUtils.java
@@ -1,4 +1,4 @@
-package me.tiary.common.util.http;
+package me.tiary.common.util.web;
 
 import jakarta.servlet.http.HttpServletRequest;
