+++ src/test/java/me/tiary/common/filter/AccessLoggingFilterTest.java
@@ -7,7 +7,7 @@ import ch.qos.logback.core.read.ListAppender;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import me.tiary.common.util.http.HttpRequestUtils;
+import me.tiary.common.util.web.HttpRequestUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
